[bug] Move convertToFileParams, convertToServerParams from replaceable license.js
# Conflicts:
#	Common/sources/utils.js
